Revert "Mobing prerender to API route using new nuxt content server API"
This reverts commit 6f34da0efae46582712bb6af9d0a33a5374a9d69.
