+++ package.json
@@ -10,7 +10,7 @@
   "devDependencies": {
     "nuxt": "3.2.3",
     "@nuxt/kit": "3.2.3",
-    "@nuxt/content": "^2.8.5",
+    "@nuxt/content": "^2.1.0",
     "@nuxt/image-edge": "^1.0.0-27809804.194d2d6",
     "@types/jsonwebtoken": "^9.0.1",
     "meilisearch": "^0.27.0",
