+++ pages/indexsearch.vue
@@ -1,0 +1,6 @@
+<template>
+    <div>
+        Indexing...
+        <IndexSearch />
+    </div>
+</template>
